refactor : 뷰 로직 분리 , useSetInputText
