+++ src/1_Page/Search/ui/SearchInput/model/useSetInputText.js
@@ -1,0 +1,9 @@
+import { useEffect } from "react";
+
+const useSetInputText = (reset, searchInputText) => {
+  useEffect(() => {
+    reset({ searchInputText: searchInputText || "" });
+  }, [searchInputText, reset]);
+};
+
+export default useSetInputText;
